requirements.txt
streamlit
pandas
PyPDF2
openai == 0.28.0
python-dotenv
fpdf2
Pillow
pytesseract
